+++ NEWS.md
@@ -1,7 +1,7 @@
 # HicAggR 0.99.3
 
 * NAMESPACE was generated with roxygen2 to define exportable functions.
-* compare_to_background: to correct the skewedness of o/e values towards long distances,
+* compareToBackground: to correct the skewedness of o/e values towards long distances,
     computation of z.scores is now calculated using residuals from a polynomial
     model that fits the background couples (log(counts)~distance).
 * SearchPairs: added option to remove self interacting bins.
@@ -14,7 +14,7 @@
 * Implemented import of O/E matrix for data in .hic format.
 * Implemented import of raw data in .h5 format.
 * Added getInfos function to get info on a hic data (.hic, cool/mcool/h5 formats).
-* Removed dependency to BSDA::z.test in compare_to_background.
+* Removed dependency to BSDA::z.test in compareToBackground.
 * Removed dependency to InteractionSet and added it as package to import in NAMESPACE to remove the PackageStartUpMessages.
 * Removed chatty package start up message and replaced it with nicer message.
 * Some BiocCheck NOTES were also takedn into consideration: changing sapply to vapply etc.
@@ -22,7 +22,7 @@
 # HicAggR 0.99.1
 
 * Corrected with Bioconductor's reviews
-* Added PrepareMtxList, import_loops, PlotAPA_byDistance & compare_to_background functions
+* Added PrepareMtxList, importLoops, plotMultiAPA & compareToBackground functions
 * Corrected bugs on seqlevels consistancy and name column for GRanges objects
 * Encapsulated small and internal functions in utilities.R
 * ExtractSubMatrix has option to remove duplicated submatrices